% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_expression.R
\name{average.expression}
\alias{average.expression}
\title{Average feature expression across clustered samples in a Seurat object using fast sparse matrix methods}
\usage{
average.expression(
  object,
  ident = "active.ident",
  assay = "default",
  slot = "counts",
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{ident}{Ident with sample clustering information (default is the active ident)}

\item{assay}{Assay to average (default is the active assay)}

\item{slot}{Slot to average (default is counts)}

\item{verbose}{Boolean or integer, show progress bar (default is TRUE)}
}
\value{
dense matrix of cluster centers
}
\description{
Average feature expression across clustered samples in a Seurat object using fast sparse matrix methods
}
