% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageexpression_sparse.R
\name{AverageExpression.Sparse}
\alias{AverageExpression.Sparse}
\title{Average cluster expression from Seurat object using fast sparse matrix methods}
\usage{
AverageExpression.Sparse(
  object,
  ident = "active.ident",
  assay = "default",
  slot = "counts",
  verbose = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{ident}{Ident to average across (default is the active Idents()). Specify any valid discrete 'object@meta.data$ slot'.}

\item{assay}{Assay to average (default is the active assay). Specify any valid assay name.}

\item{slot}{Slot to average, default is counts. Specify any valid counts slot within the assay}
}
\description{
Average cluster expression from Seurat object using fast sparse matrix methods
}
