% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_divisive.R
\name{cluster.divisive}
\alias{cluster.divisive}
\title{Fast divisive clustering of columns in a large sparse matrix}
\usage{
cluster.divisive(
  input,
  min.samples = 10,
  subsample = 10000,
  seed = NULL,
  verbose = 2,
  min.dist = NULL,
  idents.prefix = ""
)
}
\arguments{
\item{input}{A dgCMatrix matrix or a Seurat object with counts for clustering in the default assay counts slot}

\item{min.samples}{The minimum number of samples permitted in a cluster. Real valued between 2 and ncol(input), defaults to 10}

\item{subsample}{Number of samples to subsample for SVD and delta distance calculations, usually only useful for large matrices (> 10000 samples)}

\item{seed}{Random seed for subsampling (NULL by default)}

\item{verbose}{0 = no output, 1 = output for each generation, 2 = progress bar for each generation, 3 = details for each division}

\item{min.dist}{Minimum delta distance, which is a proportion of how much more similar samples are to their assigned cluster than their opponent cluster. NULL by default. Values which approximate Newman-Girvan modularity of 0 are usually between 0.01 and 0.05.}

\item{idents.prefix}{Prefix for ident names, useful to prevent later runs of cluster.divisive on the same object from overwriting previous idents}
}
\value{
If a Seurat object was input, a Seurat object with idents set to cluster IDs. If a sparse matrix was input, a list of cluster assignments for each sample.
}
\description{
Fast divisive clustering of columns in a large sparse matrix
}
