% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findclusters_divisive.R
\name{FindClusters.Divisive}
\alias{FindClusters.Divisive}
\title{Run recursive bipartite clustering by spectral decomposition with sparse matrix methods}
\usage{
FindClusters.Divisive(
  seuratObj,
  stop.modularity = NULL,
  stop.delta.dist = NULL,
  stop.min.cells = 50,
  subsample.size = 1000,
  seed = NULL,
  verbose = 1,
  shift = TRUE,
  Q.sample.size = 1000
)
}
\arguments{
\item{stop.modularity}{Optional, stopping criteria for modularity (default = NULL, modularity not used as stopping criteria). Most logical stopping criteria (if used) is 0. Newman-Girvan modularity ranges from -1 to 1.}

\item{stop.delta.dist}{Optional, stopping criteria based on distance. Faster calculation than modularity, measures the mean fold change in similarity for all cells between a child cluster center and a parent cluster center using cosine similarity.}

\item{stop.min.cells}{Required, the minimum number of cells permitted in a cluster. Real valued between 1 and ncol(A). Default 50.}

\item{subsample.size}{The maximum number of cells to sample for determining cluster centers by ssvd}

\item{seed}{Random seed for subsampling, set to NULL by default}

\item{Q.sample.size}{The maximum number of cells in each clustering for which to compute a cosine similarity graph for calculation of Q (only applicable if stop.modularity is set)}

\item{data}{Seurat object, the default assay counts slot will be used}
}
\value{
A Seurat object where `object@ident` has been updated with new cluster info. Individual ident classes are created and stashed for each generation (i.e. seuratObj@meta.data$gen1.idents). Clusters are given names as a sequence of 1 and 2 corresponding to bipartite splits, and leaf clusters names are terminated with a 0
}
\description{
Run recursive bipartite clustering by spectral decomposition with sparse matrix methods
}
