% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/main.R
\name{getDataMatrix}
\alias{getDataMatrix}
\alias{getDataMatrix.Matrix}
\alias{getDataMatrix.Seurat}
\title{Extract data matrix from Seurat object or expression matrix}
\usage{
getDataMatrix(obj, ...)

\method{getDataMatrix}{Matrix}(
  obj,
  genes = NULL,
  do_filtering = TRUE,
  min_expr = 3.5,
  min_pct = 0.2,
  center = TRUE,
  scale = FALSE,
  non_negative = TRUE,
  verbose = TRUE,
  ...
)

\method{getDataMatrix}{Seurat}(
  obj,
  assay = "RNA",
  slot = "data",
  genes = NULL,
  do_filtering = TRUE,
  min_expr = 3.5,
  min_pct = 0.2,
  center = TRUE,
  scale = FALSE,
  non_negative = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object or a \code{\link[base]{matrix}}-like object}

\item{\dots}{Further arguments passed to each method}

\item{genes}{List of variable genes to subset the matrix. If NULL, uses all genes}

\item{do_filtering}{Whether filter the gene by \code{min_expr} and \code{min_pct}.}

\item{min_expr}{The minum expression of gene expression}

\item{min_pct}{The minum precent of cell expression gene greater than \code{min_expr}.}

\item{center}{Whether to center the data matrix}

\item{scale}{Whether to scale the data matrix}

\item{non_negative}{Whether fill the negative value with zero}

\item{verbose}{Show more message}

\item{assay}{Get data matrix from this assay}

\item{slot}{Get data matrix from this slot (=layer)}
}
\value{
Returns a sparse data matrix (cells per genes), subset according to the given parameters
}
\description{
Get the gene expression matrix from a Seurat object or an expression matrix, optionally centered
and/or subset on highly variable genes
}
\examples{
library(Seurat)
data('pbmc_small')
a <- getDataMatrix(pbmc_small, genes = c('MS4A1', 'TCL1A'), do_filtering = FALSE,
                   scale = TRUE, center = TRUE, non_negative = TRUE)
}
