% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MP.R
\name{getMP}
\alias{getMP}
\title{Generate meta-program}
\usage{
getMP(
  similarity_matrix,
  nmf_programs,
  Genes_nmf_w_basis,
  initial_score = 0.2,
  min_add_score = 0.2,
  min_group_size = 5
)
}
\arguments{
\item{similarity_matrix}{jaccard similarity matrix get use \code{getSimilarity}}

\item{nmf_programs}{result get use \code{select_program_features}}

\item{Genes_nmf_w_basis}{a list of w matrix get use \code{nmf_programs}}

\item{initial_score}{the minus cutoff jaccard index for defining the first NMF program in a cluster}

\item{min_add_score}{the minus cutoff jaccard index for adding a new NMF to the forming cluster}

\item{min_group_size}{the minimal program number to consider for defining a cluster}
}
\value{
return a list with meta-program cluster member and meta-program feature set
}
\description{
Generate meta-program cluster
}
