% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSimilarity}
\alias{getSimilarity}
\title{get similarity matrix}
\usage{
getSimilarity(top_program_genes, robust_program, hclust.method = "average")
}
\arguments{
\item{top_program_genes}{a list of object get use \code{select_top_program_genes()}}

\item{robust_program}{The result of \code{robust_program()}}

\item{hclust.method}{Method to build similarity tree between individual programs}
}
\value{
return a ordered jaccard similarity matrix
}
\description{
Calculate the jaccard similarity matrix among robust program
}
