% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot similarity matrix}
\usage{
plot_similarity_matrix(
  top_program_genes,
  robust_program,
  hclust.method = "average"
)
}
\arguments{
\item{top_program_genes}{a list of object get use \code{select_top_program_genes()}}

\item{robust_program}{The result of \code{robust_program()}}

\item{hclust.method}{Method to build similarity tree between individual programs}
}
\value{
return a ggplot object
}
\description{
Plot the Jaccard similarity heatmap.
}
