% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocess.R
\name{pre_filter}
\alias{pre_filter}
\alias{pre_filter.default}
\alias{pre_filter.Matrix}
\alias{pre_filter.Seurat}
\title{Pre normalize and filter data}
\usage{
pre_filter(object, ...)

\method{pre_filter}{default}(object, min_expr = 3.5, min_pct = 0.02, verbose = TRUE, ...)

\method{pre_filter}{Matrix}(object, min_expr = 3.5, min_pct = 0.02, verbose = TRUE, ...)

\method{pre_filter}{Seurat}(
  object,
  assay = "RNA",
  min_expr = 3.5,
  min_pct = 0.02,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{min_expr}{The min expression value of a gene}

\item{min_pct}{The min percent of cells expressed a gene}

\item{verbose}{Print more info}

\item{assay}{Specific assay to get data from or set data for}
}
\value{
Return a expression matrix after normalization and filtered for \code{nmf_programs}
}
\description{
Pre normalize and filter data
}
