% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nmf.R
\name{RunNMF}
\alias{RunNMF}
\title{Non-negative matrix factorization on the average expression of clusters in a Seurat object identity class and projection of cells onto NMF coordinates}
\usage{
RunNMF(
  object,
  k = NULL,
  ident = "active.ident",
  assay = "default",
  slot = "counts",
  verbose = 1,
  rel.tol = 1e-04,
  n.threads = 0
)
}
\arguments{
\item{object}{Seurat object}

\item{k}{Decomposition rank, integer.}

\item{ident}{Ident to average across (default is the active Idents()). Specify any valid discrete 'object@meta.data$ slot'.}

\item{assay}{Assay to average across (default is the active assay). Specify any valid assay name.}

\item{slot}{Slot to average, default is counts. Specify any valid counts slot within the assay.}

\item{verbose}{0 = no tracking, 1 = progress bar relative to max.iter, 2 = iteration info. Default 2.}

\item{rel.tol}{Stop criterion, defined as the relative tolerance between two successive iterations: |e2-e1|/avg(e1,e2). Default 1e-3.}

\item{n.threads}{Number of threads/CPUs to use. Default to 0 (all cores).}

\item{return}{Seurat object with a dimensional reduction NMF slot}
}
\description{
Non-negative matrix factorization on the average expression of clusters in a Seurat object identity class and projection of cells onto NMF coordinates
}
