\name{scNMF-package}
\alias{scNMF-package}
\alias{scNMF}
\docType{package}
\title{
\packageTitle{scNMF}
}
\description{
\packageDescription{scNMF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scNMF}
\packageIndices{scNMF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scNMF}

Maintainer: \packageMaintainer{scNMF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
