% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_program_features}
\alias{select_program_features}
\title{Select NMF program features}
\usage{
select_program_features(nmf_programs_list, top_n = 50)
}
\arguments{
\item{nmf_programs_list}{a list of object get use \code{nmf_programs()}}

\item{top_n}{the number of features to select from each NMF program by importance}
}
\value{
return a list for each object with the selected top n gene of each program
}
\description{
Select top n important features for each NMF program
}
