% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_robust_nmf_program}
\alias{select_robust_nmf_program}
\title{select robust program}
\usage{
select_robust_nmf_program(
  top_program_features,
  intra_min = 0.7,
  intra_max = 0.2,
  inter_filter = T,
  inter_min = 0.2
)
}
\arguments{
\item{top_program_features}{the object obtained from \code{select_program_features}}

\item{intra_min}{minimum percents overlap with a program from the same sample (for selecting robust programs), default 0.7}

\item{intra_max}{maximum percents overlap with a program from the same sample (for removing redundant programs), default 0.2}

\item{inter_filter}{logical; indicates whether programs should be filtered based on their similarity to programs of other sample, default TRUE}

\item{inter_min}{minimum overlap with a program from another sample, default 0.2}
}
\value{
return a list of select nmf rank program for each sample. If one sample have
no robust program, it will be filtered.
}
\description{
Select robust rank program from the NMF result
}
