% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_top_program_genes}
\alias{select_top_program_genes}
\title{Select top n gene feature}
\usage{
select_top_program_genes(nmf_programs_list, top_n = 50)
}
\arguments{
\item{nmf_programs_list}{a list of object get use \code{nmf_programs()}}

\item{top_n}{the number of gene to select from each NMF program}
}
\value{
return a list for each object with the selected top n gene of each program
}
\description{
Select top n gene for each NMF program
}
