% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_divisive.R
\name{cluster.divisive}
\alias{cluster.divisive}
\title{A single instance of divisive clustering by sparse SVD}
\usage{
cluster.divisive(
  d,
  subsample.size = 5000,
  seed = NULL,
  shift = TRUE,
  dist = TRUE,
  Q = FALSE,
  Q.sample.size = 1000
)
}
\arguments{
\item{d}{A sparse dgCMatrix with genes in rows and cells in columns}

\item{subsample.size}{Number of cells to subsample for determination of cluster centers by SVD}

\item{seed}{Random seed for subsampling}

\item{shift}{Boolean, shift the clustering vector to achieve optimal bipartitioning about 0 (shift determined by the mean of both kmeans centers on the bimodal density distribution of the clustering vector)}

\item{dist}{Whether to calculate the change in distance (delta.dist) between cells and their respective child cluster center relative to the parent center}

\item{Q}{Boolean, calculate Newman-Girvan modularity}

\item{Q.sample.size}{How many cells at most to use for calculating the modularity. A square cosine similarity matrix of this size is computed. Default 1000.}
}
\value{
A list of cells in either cluster and the dip test result for bimodality between cluster centers
}
\description{
A single instance of divisive clustering by sparse SVD
}
