% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nmf_programs}
\alias{nmf_programs}
\title{Run NMF}
\usage{
nmf_programs(mat, rank = 2, method = "snmf/r", seed = 1, ...)
}
\arguments{
\item{mat}{a properly matrix data as input for \code{\link[NMF]{nmf}}, you can get it through \code{\link[scNMF]{pre_filter}}.}

\item{rank}{specification of the factorization rank. Here it must be a single numeric value.}

\item{method}{specification of the NMF algorithm.}

\item{seed}{specification of the starting point or seeding method.}

\item{...}{Other argument in \code{\link[NMF]{nmf}} function.}
}
\value{
return a list contain w and h matrix of NMF.
}
\description{
Run nonnegative matrix factorization (NMF) with matrix
}
