% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmf_project.R
\name{nnmf.project}
\alias{nnmf.project}
\title{Given W and a set of samples, calculate the optimal mapping H such that the objective mean((samples - WH)^2) is minimized}
\usage{
nnmf.project(samples, W.model, n.threads = 0)
}
\arguments{
\item{samples}{Samples to be projected onto W.model. Dense matrix.}

\item{W.model}{NMF factor model (genes x factors) for mapping to cells}

\item{n.threads}{Number of threads/CPUs to use. Default to 0 (all cores)}
}
\value{
A list of fit weights (H), an imputed counts matrix (A_imputed)
}
\description{
Given W and a set of samples, calculate the optimal mapping H such that the objective mean((samples - WH)^2) is minimized
}
