% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MP.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot similarity matrix}
\usage{
plot_similarity_matrix(
  similarity_matrix,
  hclust.method = "average",
  reverse_x = FALSE,
  reverse_y = FALSE,
  MP = NULL
)
}
\arguments{
\item{similarity_matrix}{jaccard similarity matrix get use \code{getSimilarity}}

\item{hclust.method}{cluster method for \code{\link[stats]{hclust}}}

\item{reverse_x}{reverse the order of x axis.}

\item{reverse_y}{reverse the order of y axis.}

\item{MP}{two cloumns data.frame with nmf program and meta-program cluster info,
you can get cluster info use \code{getMP}. If you supply this, the heatmap will order by meta-program cluster}
}
\value{
return a ggplot object
}
\description{
Plot the Jaccard similarity heatmap.
}
